Change "No CL&W Credit Found" error to return 0 chapel credits
This error message from the Go Gordon website indicates that the student
does not have a CL&W credit requirement or that the CL&W credit
counting has not started for the semester. In either of these cases, the
student will prefer to see the value of "0" CL&W credits instead of a
vague error message.