Fix Go Gordon reauthentication bug
Sometimes, even though the user should have to reauthenticate, it seems
that Go Gordon does not require it. This created an error where the
Mechanize browser looked for a form named 'form1' but could not find it
if Go Gordon was not asking for a reauthentication. I added error
handling that will simply skip the reauthentication if the form is not
present on the page.