Fix chapel events "next" before every weekday
Previously, a weekday in the relative date before an event would say
"next Saturday", for example, even if it was Saturday of that same week.
Now it will simply say "Saturday" if it is in the same week, and will
prepend "next" to the relative date if it is a day of the next week.